Refactor la clase Encriptar a Desencriptar: actualizar métodos, mejorar la legibilidad del código y agregar funcionalidad para enviar correos electrónicos con datos encriptados.
